Move the bulk of project resolver to Microsoft.Dnx.Tooling
